feat: refactor EditView to use new EditablePages feature
The EditView component is refactored to use the new EditablePages feature. The tests are updated to reflect the changes in the EditView component. The useAutoSave hook is refactored to use debounce and useAuthorizedUpdateContentWithStates hook.
The useAutoSave hook now returns a handler function and a state object. The state object contains the waiting, loading, error, and data states. The handler function is called when the content changes. The handler function debounces the call to authorizedUpdateContent. The authorizedUpdateContent function is called after the debounce delay. The timeout is cleared on unmount. The loading, error, and data states from useAuthorizedUpdateContentWithStates are exposed.