feat(genome-page): create SeeAllGenesChip component
This commit introduces a new `SeeAllGenesChip` component.
The component is a clickable chip that displays the label "See all".
It is used to navigate to a page containing all genes.
The chip uses the secondary color and has a small size.
It also has a margin of 0px 5px 5px 0px.