feat(genome-page): create MentionedGenes component to display genes mentioned in reference
This component displays a list of genes as clickable chips. Clicking a chip navigates the user to the corresponding gene page using the next/router. The component receives an array of gene objects, each containing an id and a name. The chips are displayed in a grid layout with a small margin.