feat(RelatedGenesDisplay.tsx): display empty GeneChip components to fill the grid
Adds GeneChipFiller component to display empty GeneChip components in the
RelatedGenesDisplay component. This ensures that the grid is always filled
with a consistent number of GeneChip components, even if there are fewer
related genes than the maximum number of components to display.
The maxCount property was added to RelatedGenesProperties to determine the
maximum number of GeneChip components to display.