feat(genome-page): create ErrorPanelC component to display errors
This commit introduces a new ErrorPanelC component that is used to display error messages in a styled panel. The component includes a title, message, error details, and a support email link. It also includes a refresh button to reload the page. This component will be used to display errors on the genome page.