refactor: replace GraphQLErrorPage with ErrorPageWrapper
This commit introduces a new `ErrorPageWrapper` component in the
`ui-dsc` package. This component wraps the existing `ErrorPage`
component from `@dictybase/ui-common` and provides additional
functionality, such as navigation to the homepage and page reload.

The `GraphQLErrorPage` component has been replaced with the new
`ErrorPageWrapper` in the following files:

- `apps/stock-center/src/pages/information/[name]/edit.tsx`
- `apps/stock-center/src/pages/information/[name]/editable.tsx`
- `apps/stock-center/src/pages/information/[name]/show.tsx`
- `apps/stock-center/src/pages/plasmids/index.tsx`
- `apps/stock-center/src/pages/strains/index.tsx`
- `packages/ui-dsc/src/catalog/EditableStrainDetailsContainer.tsx`
- `packages/ui-dsc/src/catalog/PlasmidDetailsContainer.tsx`
- `packages/ui-dsc/src/catalog/SearchPhenotypeContainer.tsx`
- `packages/ui-dsc/src/catalog/StrainDetailsContainer.tsx`

This change ensures a consistent error handling experience across the
application and provides users with helpful options for resolving
errors. Also, fetchPolicy was added to the graphql queries to improve
the user experience.