feat(publication): create ErrorPageWrapper component and use it in publication page
This commit introduces a new ErrorPageWrapper component that wraps the ErrorPage component from the @dictybase/ui-common library. The ErrorPageWrapper component is used in the publication page to display error messages. The ErrorPageWrapper component takes an error object as a prop and passes it to the ErrorPage component. The ErrorPageWrapper component also provides two functions, handleNavigateHome and handleReload, that are used to navigate to the home page and reload the page, respectively.