feat(genome-page): create SinglePublication component to display publication details
This commit introduces a new component, SinglePublication, which is responsible for displaying the details of a single publication. It includes the title, journal, publication date, authors, and related genes. The component uses Material-UI for styling and fp-ts for functional programming. It also uses date-fns for date formatting and @dictybase/ui-common for title formatting. The component is designed to be clickable, navigating the user to the publication URL when clicked. A limit is set for the number of related genes displayed, with a "See All Genes" chip appearing when the number of related genes exceeds the limit.