feat(ReferencesToolbar.tsx): add type safety to sorting value
The sorting value in the ReferencesToolbar component is now type-safe.
This was achieved by creating a new type, OrderFunctionKeys, which is
a keyof typeof orderFunctions. This type is then used to type the
setSorting function, which ensures that only valid sorting values can
be selected. Also, destructure title property to improve readability.