feat(ReferenceSortMenuItem.tsx): create component for reference sorting menu item
This commit introduces a new component, ReferenceSortMenuItem, which is responsible for rendering a menu item for sorting references. It accepts an array of tuples, where each tuple contains a string representing the sorting option and an Ord instance for comparing PublicationWithGene objects. This component will be used in the reference sorting menu to provide different sorting options to the user.